+++ apps/io-wallet-user-func/src/app/main.ts
@@ -120,7 +120,7 @@ const whitelistedFiscalCodeRepository =
 
 const pidIssuerClient = new PidIssuerClient(
   config.pidIssuer,
-  config.entityConfiguration.federationEntity.basePath.href,
+  config.entityConfiguration.federationEntity.basePathV10.href,
 );
 
 const mobileAttestationService = new MobileAttestationService(
@@ -164,7 +164,12 @@ const containerClient = blobServiceClient.getContainerClient(
 
 const certificateRepository = new CosmosDbCertificateRepository(database);
 
-const certificateIssuerAndSubject = `C=${config.walletProvider.certificate.country}, ST=${config.walletProvider.certificate.state}, L=${config.walletProvider.certificate.locality}, O=${config.entityConfiguration.federationEntity.organizationName}, CN=${config.entityConfiguration.federationEntity.basePath.hostname}`;
+const certificateV13Repository = new CosmosDbCertificateRepository(
+  database,
+  "certificates-v-1.3",
+);
+
+const certificateIssuerAndSubject = `C=${config.walletProvider.certificate.country}, ST=${config.walletProvider.certificate.state}, L=${config.walletProvider.certificate.locality}, O=${config.entityConfiguration.federationEntity.organizationName}, CN=${config.entityConfiguration.federationEntity.basePathV10.hostname}`;
 
 const statusListCatalogRepository = new CosmosDbStatusListCatalogRepository(
   database,
@@ -363,7 +368,7 @@ app.http("createWalletInstanceAttestation", {
   authLevel: "function",
   handler: CreateWalletInstanceAttestationFunction({
     assertionValidationConfig,
-    certificateRepository,
+    certificateRepository: certificateV13Repository,
     federationEntity: config.entityConfiguration.federationEntity,
     nonceRepository,
     signer: walletAttestationSigner,
@@ -381,7 +386,7 @@ app.http("createWalletUnitAttestation", {
   handler: CreateWalletUnitAttestationFunction({
     androidAttestationValidationConfig,
     assertionValidationConfig,
-    certificateRepository,
+    certificateRepository: certificateV13Repository,
     federationEntity: config.entityConfiguration.federationEntity,
     nonceRepository,
     signer: walletAttestationSigner,